moved tags to controllers
